Bug 1121571 - Read symbols from system libraries on OS X. r=jmaher
Ideally the Mozilla symbol server should have symbol files for all system libraries of all versions of all operating systems that we run on. Unfortunately, this is not the case for OS X and Linux.
This patch adds a dump_syms_mac binary that we will invoke as needed on system libraries that are i ncluded in the profile's list of shared libraries and which the official symbol server doesn't know about. We also put the resulting symbol files into a missingsymbols.zip file in the upload dir.
Obviously we should import the system library symbol files for our Talos platforms into the official Mozilla symbol server anyway. Once we've done that, this symbol dumping code won't run until we update the OS on our Talos slaves to a new version.

--HG--
extra : rebase_source : d4a18e180d692f6c3235b84eb0385524426e4ffb